powerbi/requirements.txt
simple-salesforce>=1.12.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
